pyproject.toml
[tool.poetry]
name = "sobaken-id"
version = "0.1.0"
description = ""
authors = ["Albert Bikeev <albert.bikeev@gmail.com>"]
readme = "README.md"
packages = [{ include = "data-collection" }, { include = "segmentation" }, { include = "re-identification"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}
ultralytics = "^8.3.15"
pycocotools = "^2.0.8"
pixellib = "^0.7.1"
pillow = "9.5.0"
# Deep-person-reid specific deps:
numpy = "^2.1.2"
cython = "^3.0.11"
h5py = "^3.12.1"
six = "^1.16.0"
scipy = "^1.14.1"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
tb-nightly = "^2.19.0a20241020"
future = "^1.0.0"
yacs = "^0.1.8"
gdown = "^5.2.0"
flake8 = "^7.1.1"
yapf = "^0.40.2"
isort = "4.3.21"
imageio = "^2.36.0"
chardet = "^5.2.0"
torch = {version = "^2.5.0", python = "^3.7", platform = "linux"}
# /end Deep-person-reid specific deps
imagehash = "^4.3.1"
grad-cam = "^1.5.4"
dvc = "^3.56.0"
aiohttp = "^3.10.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
